+++ thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/datalayer/bao/AnomalyFunctionManager.java
@@ -1,0 +1,5 @@
+package com.linkedin.thirdeye.datalayer.bao;
+
+public class AnomalyFunctionManager {
+
+}
