+++ thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/datalayer/dao/AbstractMappingDAO.java
@@ -1,0 +1,7 @@
+package com.linkedin.thirdeye.datalayer.dao;
+
+import com.linkedin.thirdeye.datalayer.entity.AbstractMappingEntity;
+
+public abstract class AbstractMappingDAO<E extends AbstractMappingEntity> {
+
+}
