+++ thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/datalayer/dao/AnomalyFeedbackDAO.java
@@ -1,0 +1,10 @@
+package com.linkedin.thirdeye.datalayer.dao;
+
+import com.linkedin.thirdeye.db.entity.AnomalyFeedback;
+
+public class AnomalyFeedbackDAO extends AbstractBaseDAO<AnomalyFeedback> {
+
+  public AnomalyFeedbackDAO() {
+    super(AnomalyFeedback.class);
+  }
+}
