+++ thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/datalayer/dto/AnomalyFunctionDTO.java
@@ -1,0 +1,5 @@
+package com.linkedin.thirdeye.datalayer.dto;
+
+public class AnomalyFunctionDTO {
+
+}
