+++ thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/datalayer/entity/AbstractJsonEntity.java
@@ -1,0 +1,21 @@
+package com.linkedin.thirdeye.datalayer.entity;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+public abstract class AbstractJsonEntity extends AbstractEntity {
+  /**
+   * keep it in threadlocal if obj-mapper configuration is needed to change
+   */
+  final static ObjectMapper OBJECT_MAPPER = new ObjectMapper();
+  // TODO: make this DTO typed
+
+  protected String jsonVal;
+
+  public String getJsonVal() {
+    return jsonVal;
+  }
+
+  public void setJsonVal(String jsonVal) {
+    this.jsonVal = jsonVal;
+  }
+}
