+++ thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/datalayer/entity/AnomalyFeedback.java
@@ -1,0 +1,13 @@
+package com.linkedin.thirdeye.datalayer.entity;
+
+public class AnomalyFeedback extends AbstractJsonEntity {
+  String type;
+
+  public String getType() {
+    return type;
+  }
+
+  public void setType(String type) {
+    this.type = type;
+  }
+}
