+++ thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/datalayer/entity/AnomalyFunctionMergeConfigMapping.java
@@ -1,0 +1,12 @@
+package com.linkedin.thirdeye.datalayer.entity;
+
+import java.util.List;
+
+public class AnomalyFunctionMergeConfigMapping
+    extends AbstractMappingEntity<AnomalyFunction, AnomalyMergeConfig> {
+
+  public AnomalyFunctionMergeConfigMapping(List<Long> anomalyFunctionIds,
+      List<Long> anomalyMergeConfigIds) {
+    super(anomalyFunctionIds, anomalyMergeConfigIds);
+  }
+}
