+++ thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/datalayer/entity/AnomalyRawMergedResultMapping.java
@@ -1,0 +1,4 @@
+package com.linkedin.thirdeye.datalayer.entity;
+
+public class AnomalyRawMergedResultMapping {
+}
