+++ thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/datalayer/entity/EmailFunctionMapping.java
@@ -1,0 +1,11 @@
+package com.linkedin.thirdeye.datalayer.entity;
+
+import java.util.List;
+
+public class EmailFunctionMapping
+    extends AbstractMappingEntity<EmailConfiguration, AnomalyFunction> {
+
+  public EmailFunctionMapping(List<Long> emailConfigIds, List<Long> anomalyFunctionIds) {
+    super(emailConfigIds, anomalyFunctionIds);
+  }
+}
