+++ thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/datalayer/entity/WebappConfig.java
@@ -1,0 +1,31 @@
+package com.linkedin.thirdeye.datalayer.entity;
+
+public class WebappConfig extends AbstractJsonEntity {
+  String name;
+  String collection;
+  String type;
+
+  public String getCollection() {
+    return collection;
+  }
+
+  public void setCollection(String collection) {
+    this.collection = collection;
+  }
+
+  public String getName() {
+    return name;
+  }
+
+  public void setName(String name) {
+    this.name = name;
+  }
+
+  public String getType() {
+    return type;
+  }
+
+  public void setType(String type) {
+    this.type = type;
+  }
+}
