--- thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/dbi/dao/AbstractMappingDAO.java
@@ -1,7 +1,0 @@
-package com.linkedin.thirdeye.dbi.dao;
-
-import com.linkedin.thirdeye.dbi.entity.base.AbstractMappingEntity;
-
-public abstract class AbstractMappingDAO<E extends AbstractMappingEntity> {
-
-}
