--- thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/dbi/dao/AnomalyFeedbackDAO.java
@@ -1,10 +1,0 @@
-package com.linkedin.thirdeye.dbi.dao;
-
-import com.linkedin.thirdeye.db.entity.AnomalyFeedback;
-
-public class AnomalyFeedbackDAO extends AbstractBaseDAO<AnomalyFeedback> {
-
-  public AnomalyFeedbackDAO() {
-    super(AnomalyFeedback.class);
-  }
-}
