--- thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/dbi/entity/data/AnomalyMergedResult.java
@@ -1,87 +1,0 @@
-package com.linkedin.thirdeye.dbi.entity.data;
-
-import com.linkedin.thirdeye.dbi.entity.base.AbstractJsonEntity;
-
-public class AnomalyMergedResult extends AbstractJsonEntity {
-  long anomalyFunctionId;
-  long anomalyFeedbackId;
-  long metricId;
-  long startTime;
-  long endTime;
-  String collection;
-  String metric;
-  String dimensionValue;
-  boolean notified;
-
-  public long getAnomalyFeedbackId() {
-    return anomalyFeedbackId;
-  }
-
-  public void setAnomalyFeedbackId(long anomalyFeedbackId) {
-    this.anomalyFeedbackId = anomalyFeedbackId;
-  }
-
-  public long getAnomalyFunctionId() {
-    return anomalyFunctionId;
-  }
-
-  public void setAnomalyFunctionId(long anomalyFunctionId) {
-    this.anomalyFunctionId = anomalyFunctionId;
-  }
-
-  public String getCollection() {
-    return collection;
-  }
-
-  public void setCollection(String collection) {
-    this.collection = collection;
-  }
-
-  public String getDimensionValue() {
-    return dimensionValue;
-  }
-
-  public void setDimensionValue(String dimensionValue) {
-    this.dimensionValue = dimensionValue;
-  }
-
-  public long getEndTime() {
-    return endTime;
-  }
-
-  public void setEndTime(long endTime) {
-    this.endTime = endTime;
-  }
-
-  public String getMetric() {
-    return metric;
-  }
-
-  public void setMetric(String metric) {
-    this.metric = metric;
-  }
-
-  public long getMetricId() {
-    return metricId;
-  }
-
-  public void setMetricId(long metricId) {
-    this.metricId = metricId;
-  }
-
-  public long getStartTime() {
-    return startTime;
-  }
-
-  public void setStartTime(long startTime) {
-    this.startTime = startTime;
-  }
-
-  public boolean isNotified() {
-    return notified;
-  }
-
-  public void setNotified(boolean notified) {
-    this.notified = notified;
-  }
-}
