--- thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/dbi/entity/data/IngraphMetricConfig.java
@@ -1,24 +1,0 @@
-package com.linkedin.thirdeye.dbi.entity.data;
-
-import com.linkedin.thirdeye.dbi.entity.base.AbstractJsonEntity;
-
-public class IngraphMetricConfig extends AbstractJsonEntity {
-  String name;
-  String alias;
-
-  public String getAlias() {
-    return alias;
-  }
-
-  public void setAlias(String alias) {
-    this.alias = alias;
-  }
-
-  public String getName() {
-    return name;
-  }
-
-  public void setName(String name) {
-    this.name = name;
-  }
-}
