--- thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/dbi/entity/data/WebappConfig.java
@@ -1,33 +1,0 @@
-package com.linkedin.thirdeye.dbi.entity.data;
-
-import com.linkedin.thirdeye.dbi.entity.base.AbstractJsonEntity;
-
-public class WebappConfig extends AbstractJsonEntity {
-  String name;
-  String collection;
-  String type;
-
-  public String getCollection() {
-    return collection;
-  }
-
-  public void setCollection(String collection) {
-    this.collection = collection;
-  }
-
-  public String getName() {
-    return name;
-  }
-
-  public void setName(String name) {
-    this.name = name;
-  }
-
-  public String getType() {
-    return type;
-  }
-
-  public void setType(String type) {
-    this.type = type;
-  }
-}
