--- thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/dbi/entity/full/AnomalyFeedbackComplete.java
@@ -1,7 +1,0 @@
-package com.linkedin.thirdeye.dbi.entity.full;
-
-import com.linkedin.thirdeye.dbi.entity.base.AbstractEntity;
-
-public class AnomalyFeedbackComplete extends AbstractEntity {
-  // full entity with all fields
-}
