--- thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/dbi/entity/mapping/AnomalyRawMergedResultMapping.java
@@ -1,4 +1,0 @@
-package com.linkedin.thirdeye.dbi.entity.mapping;
-
-public class AnomalyRawMergedResultMapping {
-}
